requirements.txt
sanic
Pillow
tortoise-orm[asyncpg]